#2231 move java rpc handler examples to a package
