+++ example/main-java/src/main/java/org/finos/vuu/module/JavaExampleModule.java
@@ -13,11 +13,11 @@ import org.finos.vuu.core.table.DefaultColumn;
 import org.finos.vuu.core.table.TableContainer;
 import org.finos.vuu.net.rpc.AllowAllRpcPermissionChecker$;
 import org.finos.vuu.net.rpc.RpcHandler;
-import org.finos.vuu.person.DeleteRecordRpcHandler;
-import org.finos.vuu.person.DeleteRecordRpcHandlerIF;
-import org.finos.vuu.person.EditPersonRecordRpcHandler;
-import org.finos.vuu.person.PersonRpcHandler;
-import org.finos.vuu.person.UpdateRecordRpcHandler;
+import org.finos.vuu.person.rpc.DeleteRecordRpcHandler;
+import org.finos.vuu.person.rpc.DeleteRecordRpcHandlerIF;
+import org.finos.vuu.person.rpc.EditPersonRecordRpcHandler;
+import org.finos.vuu.person.rpc.PersonRpcHandler;
+import org.finos.vuu.person.rpc.UpdateRecordRpcHandler;
 import org.finos.vuu.person.auto.AutoMappedPersonProvider;
 import org.finos.vuu.person.auto.EntitySchema;
 import org.finos.vuu.person.datasource.PersonStore;
